Changelog - handle network error with Snackbar UI - added CoinHeaderAdapter.kt for Top ranking items - implement getItemViewType() to identify viewtypes from ConcatAdapter - implement Recyclerview's GridLayoutManager due to switch between vertical and horizontal list - implement ranking logic for header - implement new ViewHolder for header - modify FetchAndSaveDataUseCase.kt - added global variable for checking isNetworkAvailable - added global function for checking listenForNetworkConnection() - refactor naming success to done at Event.kt , EventHandler.kt, ExtensionObservers.kt - modify CoinDao.kt to retrieve ordered list - added new logic flow isConnectedToNetwork() at FetchAndSaveDataUseCase.kt - added Context.isConnectedToNetwork() at Extensions.kt - register Context with DI at ApiModule.kt - added NoInternetConnectionError at CustomError.kt - added styles.xml with colors.xml

+++ app/src/main/java/com/htetwill/coinranking/data/usecase/FetchAndSaveDataUseCase.kt
@@ -1,29 +1,43 @@
 package com.htetwill.coinranking.data.usecase
 
+import android.content.Context
 import android.util.Log
+import com.htetwill.coinranking.data.event.DoneEvent
 import com.htetwill.coinranking.data.event.ErrorEvent
 import com.htetwill.coinranking.data.event.Event
 import com.htetwill.coinranking.data.event.LoadingEvent
-import com.htetwill.coinranking.data.event.SuccessEvent
 import com.htetwill.coinranking.data.repository.ApplicationRepository
+import com.htetwill.coinranking.data.util.isConnectedToNetwork
+import com.htetwill.coinranking.error.NoInternetConnectionError
 import javax.inject.Inject
 
 class FetchAndSaveDataUseCase @Inject constructor(
-    private val mApplicationRepository: ApplicationRepository) {
-    suspend fun invoke(value: Unit): Event<Unit> {
-        when(val fetchPostResponse = mApplicationRepository.fetchData()) {
-            is SuccessEvent -> {
-                mApplicationRepository.deleteData()
-                fetchPostResponse.data.run {
-                    when(val saveDataResponse = mApplicationRepository.saveData(this!!.data!!.coins)){
-                        is ErrorEvent -> return ErrorEvent(saveDataResponse.error)
-                        is LoadingEvent -> Log.wtf("fetchPostResponse","LoadingEvent")
-                        is SuccessEvent -> Log.wtf("fetchPostResponse","SuccessEvent")
+    private val mApplicationRepository: ApplicationRepository,
+    private val mContext: Context
+) {
+    suspend fun invoke(): Event<Unit> {
+        return try {
+            if(mContext.isConnectedToNetwork()) {
+                when(val fetchPostResponse = mApplicationRepository.fetchData()) {
+                    is DoneEvent -> {
+                        mApplicationRepository.deleteData()
+                        fetchPostResponse.data.run {
+                            when(val saveDataResponse =
+                                mApplicationRepository.saveData(this!!.data!!.coins)) {
+                                is ErrorEvent -> return ErrorEvent(saveDataResponse.error)
+                                is LoadingEvent -> Log.wtf("fetchPostResponse", "LoadingEvent")
+                                is DoneEvent -> Log.wtf("fetchPostResponse", "DoneEvent")
+                            }
+                        }
                     }
+                    is ErrorEvent -> return ErrorEvent(fetchPostResponse.error)
                 }
+                DoneEvent(Unit)
+            }else{
+                ErrorEvent(NoInternetConnectionError)
             }
-            is ErrorEvent -> return ErrorEvent(fetchPostResponse.error)
+        } catch(e: Exception) {
+            return DoneEvent(Unit)
         }
-        return SuccessEvent(Unit)
     }
 }
